(#1) Starting to look like a real allocator
This commit modifies the actual allocator::Trollocator to start looking like an actual, extremely rudimentary allocator. It is still completely unfit for GlobalAlloc and unable to expand past 4096 bytes (which means at most 100 or so blocks), and does nothing about fragmentation, but it can at least allocate a block at this point. 🤗