(#5) Add debugging code
Adds code to debug the segmentation fault, because the stack gets horribly corrupted when using GDB to backtrace.

It looks like coalesce might be failing and the function to retrieve heap start was previously wrong. I don't know if the coalescence has to do with the segfault though. Code that is used inside the allocator is calling the allocator, I think, which is probably problematic.